changed server2 to server
