+++ docker-compose.yml
@@ -89,26 +89,13 @@ services:
       - cache_volume:/cache/mmaction
       - ./utils:/shared_classes/utils
 
-#  producer:
-#    build: .
-#    container_name: "producer_vision"
-#    ports:
-#      - "5002:5001"
-#    command: bash -c "python3 producer.py -u ${REDIS_URL} --inputFps ${INPUT_FPS} --outputFps ${OUTPUT_FPS} ${INPUT_FILE} --webcam ${CAMERA} & python3 server.py -u ${REDIS_URL} --trackletLength ${TRACKLET_LENGTH} --score ${SCORE}"
-#    depends_on:
-#      redis:
-#        condition: service_healthy
-#    profiles:
-#      - tracking
-#    environment:
-#      - MODEL_CONF=${MODEL_CONF}
 
   producer:
     build: .
     container_name: "producer_vision"
     ports:
       - "5002:5001"
-    command: bash -c "python3 producer.py -u ${REDIS_URL} --inputFps ${INPUT_FPS} --outputFps ${OUTPUT_FPS} ${INPUT_FILE} --webcam ${CAMERA} & python3 server2.py -u ${REDIS_URL} --trackletLength ${TRACKLET_LENGTH} --score ${SCORE} --actionrecThreshold ${ACTION_REC_THRESHOLD} --videoThreshold ${VIDEO_THRESHOLD}"
+    command: bash -c "python3 producer.py -u ${REDIS_URL} --inputFps ${INPUT_FPS} --outputFps ${OUTPUT_FPS} ${INPUT_FILE} --webcam ${CAMERA} & python3 server.py -u ${REDIS_URL} --trackletLength ${TRACKLET_LENGTH} --score ${SCORE} --actionrecThreshold ${ACTION_REC_THRESHOLD} --videoThreshold ${VIDEO_THRESHOLD}"
     depends_on:
       redis:
         condition: service_healthy
